fix: move Move PlanCardProps to components
