--- apps/web/types/planCard.ts
@@ -1,9 +1,0 @@
-export interface PlanCardProps {
-  title: string;
-  price: string;
-  desc1: string;
-  desc2: string;
-  features: string[];
-  cta: string;
-  highlight?: boolean;
-}
